requirements.txt
langchain-core>=0.3.0,<0.4.0
langchain>=0.3.0,<0.4.0
langchain-groq>=0.2.0,<0.3.0
streamlit>=1.41.0
python-dotenv>=1.0.0
python-docx>=1.1.0
pydantic>=2.0.0,<3.0.0
PyGithub>=2.1.0
